horizonclient: Make operation_request examples fully qualified
