+++ clients/horizonclient/examples_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/stellar/go/clients/horizonclient"
 	hProtocol "github.com/stellar/go/protocols/horizon"
 	"github.com/stellar/go/protocols/horizon/effects"
+	"github.com/stellar/go/protocols/horizon/operations"
 )
 
 func ExampleClient_NextAssetsPage() {
@@ -128,6 +129,40 @@ func ExampleClient_NextOffersPage() {
 	for recordsFound {
 		// next page
 		nextPage, err := client.NextOffersPage(page)
+		if err != nil {
+			fmt.Println(err)
+			return
+		}
+
+		page = nextPage
+		if len(nextPage.Embedded.Records) == 0 {
+			recordsFound = false
+		}
+		fmt.Println(nextPage)
+	}
+}
+
+func ExampleClient_NextOperationsPage() {
+	client := horizonclient.DefaultPublicNetClient
+	// all operations
+	operationRequest := horizonclient.OperationRequest{Limit: 20}
+	ops, err := client.Operations(operationRequest)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	fmt.Print(ops)
+
+	// get next pages.
+	recordsFound := false
+	if len(ops.Embedded.Records) > 0 {
+		recordsFound = true
+	}
+	page := ops
+	// get the next page of records if recordsFound is true
+	for recordsFound {
+		// next page
+		nextPage, err := client.NextOperationsPage(page)
 		if err != nil {
 			fmt.Println(err)
 			return
@@ -272,6 +307,40 @@ func ExampleClient_PrevOffersPage() {
 	}
 }
 
+func ExampleClient_PrevOperationsPage() {
+	client := horizonclient.DefaultPublicNetClient
+	// all operations
+	operationRequest := horizonclient.OperationRequest{Limit: 20}
+	ops, err := client.Operations(operationRequest)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	fmt.Print(ops)
+
+	// get prev pages.
+	recordsFound := false
+	if len(ops.Embedded.Records) > 0 {
+		recordsFound = true
+	}
+	page := ops
+	// get the prev page of records if recordsFound is true
+	for recordsFound {
+		// prev page
+		prevPage, err := client.PrevOperationsPage(page)
+		if err != nil {
+			fmt.Println(err)
+			return
+		}
+
+		page = prevPage
+		if len(prevPage.Embedded.Records) == 0 {
+			recordsFound = false
+		}
+		fmt.Println(prevPage)
+	}
+}
+
 func ExampleClient_StreamEffects() {
 	client := horizonclient.DefaultTestNetClient
 	// all effects
@@ -330,6 +399,48 @@ func ExampleClient_StreamOffers() {
 		fmt.Println(offer)
 	}
 	err := client.StreamOffers(ctx, offerRequest, printHandler)
+	if err != nil {
+		fmt.Println(err)
+	}
+}
+
+func ExampleClient_StreamOperations() {
+	client := horizonclient.DefaultTestNetClient
+	// operations for an account
+	opRequest := horizonclient.OperationRequest{ForAccount: "GAIH3ULLFQ4DGSECF2AR555KZ4KNDGEKN4AFI4SU2M7B43MGK3QJZNSR", Cursor: "760209215489"}
+
+	ctx, cancel := context.WithCancel(context.Background())
+	go func() {
+		// Stop streaming after 60 seconds.
+		time.Sleep(60 * time.Second)
+		cancel()
+	}()
+
+	printHandler := func(op operations.Operation) {
+		fmt.Println(op)
+	}
+	err := client.StreamOperations(ctx, opRequest, printHandler)
+	if err != nil {
+		fmt.Println(err)
+	}
+}
+
+func ExampleClient_StreamPayments() {
+	client := horizonclient.DefaultTestNetClient
+	// all payments
+	opRequest := horizonclient.OperationRequest{Cursor: "760209215489"}
+
+	ctx, cancel := context.WithCancel(context.Background())
+	go func() {
+		// Stop streaming after 60 seconds.
+		time.Sleep(60 * time.Second)
+		cancel()
+	}()
+
+	printHandler := func(op operations.Operation) {
+		fmt.Println(op)
+	}
+	err := client.StreamPayments(ctx, opRequest, printHandler)
 	if err != nil {
 		fmt.Println(err)
 	}
